Version 1 for First prototype test
2 test type:
Test Type 1: Give all the nodes for the shortest path
Test Type 2: Give all the nodes for the shortest path, given your
current location, give you the bearings you need to turn and distance
need to travel to the next node.

Supposedly fits all the criteria but still in testing phase. ( Its a
pain in the ass to test this thing)